backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
python-decouple>=3.8
PyJWT>=2.8.0
psycopg2-binary>=2.9.10
Pillow>=11.3.0
django-cors-headers>=4.7.0
dj-database-url
gunicorn
dj-database-url
drf-yasg
whitenoise
